+++ tests/unit/controllers/FormsControllerTest.php
@@ -11,7 +11,6 @@ use putyourlightson\campaign\records\PendingContactRecord;
 use putyourlightson\campaigntests\fixtures\ContactsFixture;
 use putyourlightson\campaigntests\fixtures\MailingListsFixture;
 use putyourlightson\campaigntests\fixtures\PendingContactsFixture;
-use putyourlightson\campaigntests\unit\controllers\BaseControllerTest;
 use yii\web\NotFoundHttpException;
 
 /**